requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
databases==0.8.0
docopt==0.6.2
ecdsa==0.18.0
fastapi==0.104.1
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
packaging==23.2
passlib==1.7.4
pipreqs==0.4.13
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
yarg==0.1.9
python-multipart
